+++ app/src/main/java/it/univpm/filmaccio/data/models/Person.kt
@@ -3,7 +3,7 @@ package it.univpm.filmaccio.data.models
 import com.google.gson.annotations.SerializedName
 
 data class Person(
-    @SerializedName("id") val id: Int,
+    @SerializedName("id") val id: Long,
     @SerializedName("name") val name: String,
     @SerializedName("profile_path") val profilePath: String?,
     @SerializedName("biography") val biography: String,
